move testing_utils to conftest to appease stickler
