+++ tests/test_core.py
@@ -7,7 +7,7 @@ import pandas.util.testing as pdt
 
 from message_ix import Scenario
 from numpy import testing as npt
-from testing_utils import test_mp
+from conftest import test_mp
 
 msg_args = ('canning problem (MESSAGE scheme)', 'standard')
 msg_multiyear_args = ('canning problem (MESSAGE scheme)', 'multi-year')
